Implement missing functions of SteamMatch
The calculated variable encodingHelper was never implemented (a fatal error was in place because at the time of writing it, SteamMatchEncodingHelper didn't exist), so this was fixed.  aggregateMatchData is a required function that throws a fatal error in the super class implementation.  Since we don't do any aggregation, an empty stub was put in place to override the super class implementation (so no fatal error is thrown).