Add new enums for clarity
Some of the SteamMatch properties use ints to represent different states that are mutually exclusive.  To aid with code clarity, this commit creates enums to use instead of raw values.  This commit also updates the UpdateType enum with autonomous and teleop elements so those can be used as valid identifier to update portions of match properties

Issues: #24