Fully Implement AutoViewController
The implementation of AutonomousViewController has some values that persist between changes to the UI, but this commit fully implements the UI and uses a SteamMatch instance to persist the state.  The currentMatch's data is loaded when the view will appear and the currentMatch is correctly updated when this view has disappeared.  This functionality has been tested and works.

Issues: #24