Revert "Updated results storyboard and received problems when trying to look at the data."
I suppose I should have explained more about the reasoning for the Results Storyboard:

There is a custom container array controller that takes an array of view controllers and cascades them on top of each other.
This is the view controller that is shown in the Main Storyboard.  When this segue is used, we instantiate the proper view controllers from the Results Storyboard and add them to this container.  Thus, there is no need to delete this view from the main storyboard, we just have to update the views in the Results Storyboard.  This commit reverts the changes made.

This reverts commit 8bde5e4b21cb701862f56d7942bf7aebe3e4c0b2.