Fix bug with alliance selection controls
alliance selection code was not properly selecting the right button.  This caused an issue with the UI feedback and the fix for that caused a misrepresentation of the alliance type within the Match data structure.  The underlying implementation was changed, fixing the UI issue and the data structure misrepresentation.

Issues: #29