Removed references to SteamMatch in MatchStore
This commit uses Swift's type parameters to pass in a dynamic type that conforms to the CsvDataProvider protocol.  This allows a dynamic type to be passed in a a dynamic CSV Header to be generated.  Due to this function definition changes, the necessary changes have been made for callers of this function.

Issues: #44