Fix rotation bug in Results Views
The CustomContainerArrayView does not correctly respond to rotation changes.  The animations and positions of the Result views get messed up and you have to select another match and reselect this match to get fix it.  This commit adds the correct orientation changes so rotating the device correctly repositions the Results views.

Issues: #28