Add better orientation locking
The orientation locking is now fixed with a new method to lock/unlock the rotations.  The extension on UINavigationController has been changed to a subclass so only certain UINavigationControllers will have this effect.  The previous commit also changes the transitions of the master views to push segues to fix the rotation locks.  The Choose Event view was also changed to a form sheet instead of page sheet so it is more centered in the view.

Issues: #54