Add basic connection through Multipeer Connectivity
This commit adds a basic protocol to connect two devices through the Multipeer Connectivity Framework.
There is no data that is sent, but this shows a proof of concept in terms on connection.  The following steps
are used to test the connection: 1) One device (Device A) switches advertising to the On position; 2) The other
device (Device B) switches the browsing switch to the On position; 3) Device A will receive an invitation and a
prompt to accept. 4) After accepting, the connection will be made (looking at the logs)

The next step is to test the data transfer, first with a string, then with a serialized object (an instance of SteamMatch).

Issues: #81