Add proof of concept for message sending
Messages can now be sent between connected devices.  Once the devices are connected, a "ping" button is enabled, which sends a simple string to all connected devices.  If the message is received on a connected device, an alert shows up to indicate the ping was received (it also shows the display name of the sender).  There is some delay to the disabling of this feature, so undefined behavior can occur when the session is disconnected, but a ping is still called.  This can be fixed in later commits.  The next step is testing a serialized SteamMatch instance.

Issues: #81