Replace Version String with Enum
This commit replaces the version string representation with an enum to
allow better typing and clearer code. The current version (0.1.0) was
added along with the next minor and patch numbers. A version key was
also added to MatchTransferData in case this struct changes independent
of the discovery info struct.

Additionally, this commit updates project files to using Xcode 8.3 and
suppresses some compiler warnings

Issue: #81