Reimplement Event Transitions
The following commit reimplements event transitions in order to correct
an existing bug with Xcode and the SourceKitService. Computationally
complex assignments end up stalling the SourceKitService during indexing.
This causes Xcode to halt the cpu usage to skyrocket. Instead, one
closure is used to setup the event mapping which allows Xcode to index
properly.

Issue: #81