Implement Event Handlers on ServiceStore
This implements event handlers for the event transitions defined for the
ServiceStateMachine. The events include calls to internal elements as
well as calls to the delegate for handling UI updates.

The delegate protocol has been updated include functions for handling the
UI updates. These have been added to the view controller, but have not
been implemented yet. The protocol is not final yet, and things may
change depending on the similarities of the implementation of these
functions.

Issue: #81