Restructure Service State Events and Transitions
This commit removes unnecessary state such as the complete and error
states. These are not necessary as the event handler for transitioning to
these events would automatically pass through to the "not ready" state
immediately. They have been removed and the transitions have been adjusted
accordingly.

This commit also adds the much needed "Reset" event which safely resets
all states back to "Not Ready." This is used to stop the current mode
of advertising or browsing and easily transition to the other mode.

The last major change included in this commit is the update of the
ServiceStoreDelegate. Since most of the "handle..." functions would
amount to similar work, they have been condensed to one function that
passes all necessary information about a transition. The delegate can
then more easily respond to changes and they are not necessarily tied to
UI updates. This new function has been implemented in the DataTransferVC
and it has been made a delegate of the ServiceStore.

> NOTE: The implementation only prints out information, so further work will
> be necessary to full implement the function.

There are number of minor changes included as well:
- ServiceStore.state has been renamed to .sessionState to be more specific
- ServiceStoreDelegate is restricted to classes in order to make it a weak type
- Detailed descriptions of the states have been added to the state machine

Issue: #81