Add Selected Check to Match Data Model
This commit changes the data structure of the match model in order to
support internal data about a match. The first use of this is a check
to make sure the match was selected for data transfer. Since this will
affect all matches, it has been implemented in a way that applies to all
Match subclasses for free.

Issue: #81